Moves application back to root package and adds archetype properties file
