+++ java/Dockerfile
@@ -27,6 +27,6 @@ ARG DEPENDENCY=/workspace/app/target/dependency
 COPY --from=build ${DEPENDENCY}/BOOT-INF/lib /app/lib
 COPY --from=build ${DEPENDENCY}/META-INF /app/META-INF
 COPY --from=build ${DEPENDENCY}/BOOT-INF/classes /app
-ENTRYPOINT ["java","-cp","app:app/lib/*","com.amido.stacks.app.Application"]
+ENTRYPOINT ["java","-cp","app:app/lib/*","com.amido.stacks.Application"]
 
 EXPOSE 9000
