+++ java/archetype.properties
@@ -1,0 +1,10 @@
+# these are standard properties
+package=com.amido.stacks
+groupId=com.amido
+artifactId=stacks-api-cqrs-events
+version=1.0.7-SNAPSHOT
+archetype.filteredExtensions=java,xml,md,yml,dockerfile
+
+# here we add our personalized properties
+defaultClassPrefix=ArchetypeExample
+defaultVariablePrefix=archetypeVariable
