+++ java/build_archetype.sh
@@ -1,0 +1,2 @@
+mvn clean archetype:create-from-project -DpropertyFile=archetype.properties
+mvn install -f target/generated-sources/archetype/pom.xml
