+++ java/pom.xml
@@ -11,7 +11,7 @@
 
     <groupId>com.amido</groupId>
     <artifactId>stacks-api-cqrs-events</artifactId>
-    <version>1.0.0-SNAPSHOT</version>
+    <version>1.0.7-SNAPSHOT</version>
 
     <name>stacks-api-cqrs-events</name>
     <description>Demo project for Java with CQRS and event messaging using Azure ServiceBus</description>
@@ -400,13 +400,13 @@
                     </execution>
                 </executions>
             </plugin>
-            <plugin>
-                <groupId>org.springframework.boot</groupId>
-                <artifactId>spring-boot-maven-plugin</artifactId>
-                <configuration>
-                    <mainClass>com.amido.stacks.app.Application</mainClass>
-                </configuration>
-            </plugin>
+<!--            <plugin>-->
+<!--                <groupId>org.springframework.boot</groupId>-->
+<!--                <artifactId>spring-boot-maven-plugin</artifactId>-->
+<!--                <configuration>-->
+<!--                    <mainClass>com.amido.stacks.Application</mainClass>-->
+<!--                </configuration>-->
+<!--            </plugin>-->
             <plugin>
                 <groupId>org.jacoco</groupId>
                 <artifactId>jacoco-maven-plugin</artifactId>
