+++ java/src/main/java/com/amido/stacks/Application.java
@@ -1,4 +1,4 @@
-package com.amido.stacks.app;
+package com.amido.stacks;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
