+++ java/src/test/java/com/amido/stacks/menu/api/v1/impl/CreateCategoryControllerImplTest.java
@@ -11,7 +11,7 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static org.springframework.http.HttpStatus.BAD_REQUEST;
 
-import com.amido.stacks.app.Application;
+import com.amido.stacks.Application;
 import com.amido.stacks.core.api.dto.ErrorResponse;
 import com.amido.stacks.menu.api.v1.dto.request.CreateCategoryRequest;
 import com.amido.stacks.menu.api.v1.dto.response.ResourceCreatedResponse;
