+++ java/src/test/java/com/amido/stacks/menu/api/v1/impl/CreateItemControllerImplTest.java
@@ -15,7 +15,7 @@ import static org.springframework.http.HttpStatus.BAD_REQUEST;
 import static org.springframework.http.HttpStatus.CONFLICT;
 import static org.springframework.http.HttpStatus.NOT_FOUND;
 
-import com.amido.stacks.app.Application;
+import com.amido.stacks.Application;
 import com.amido.stacks.core.api.dto.ErrorResponse;
 import com.amido.stacks.menu.api.v1.dto.request.CreateItemRequest;
 import com.amido.stacks.menu.api.v1.dto.response.ResourceCreatedResponse;
