+++ java/src/test/java/com/amido/stacks/menu/api/v1/impl/CreateMenuControllerImplTest.java
@@ -8,7 +8,7 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.when;
 
-import com.amido.stacks.app.Application;
+import com.amido.stacks.Application;
 import com.amido.stacks.core.api.dto.ErrorResponse;
 import com.amido.stacks.menu.api.v1.dto.request.CreateMenuRequest;
 import com.amido.stacks.menu.api.v1.dto.response.ResourceCreatedResponse;
